Ensure `@stdlib` is on `LOAD_PATH` for `.pkg/` hooks
We need to be able to load the standard libraries such as `TOML`,
`Artifacts`, etc... when running our `.pkg` hooks such as
`.pkg/select_artifacts.jl`.  While these are usually available, they are
not available during tests, so let's ensure that `@stdlib` is always on
the load path.

To trigger the issue, use the following code on Julia v1.8 or later, as
only Julia v1.8+ is compatible with an `LLVM_jll` new enough to use the
`.pkg` hooks.

```
import Pkg

mktempdir() do dir
    cd(dir)
    Pkg.activate(dir)
    Pkg.generate("SomePkg")
    mkdir("SomePkg/test")
    touch("SomePkg/test/runtests.jl")
    Pkg.activate("SomePkg/test") do
        Pkg.add("LLVM_jll")
    end
    Pkg.activate("SomePkg") do
        Pkg.test()
    end
end
```